LayoutTests/inspector-protocol/layer-tree-generated-content.html

<html>
<head>
<script type="text/javascript" src="../http/tests/inspector-protocol/resources/protocol-test.js"></script>
<script type="text/javascript">

function test()
{

    getDocument();
    
    function getDocument()
    {
        // We must first get the document so that later on we may get sensible nodeIds.
        step({
            name: "Get the Document",
            command: "DOM.getDocument",
            parameters: {},
            callback: enableLayerTreeAgent
        });
    };
    
    function enableLayerTreeAgent(result)
    {
        step({
            name: "Enable the LayerTree agent",
            command: "LayerTree.enable",
            parameters: {},
            callback: getLayerTree
        });
    };

    function getLayerTree(result)
    {
        step({
            name: "Get the layer tree",
            command: "LayerTree.getLayerTree",
            parameters: {},
            callback: gotLayerTree
        });
    };

    var layerCount;

    function gotLayerTree(result)
    {
        var flatTree = flattenedLayerTree(result.layerTree);
        var layerIds = Object.keys(flatTree);
        
        layerCount = layerIds.length;

        logTestName("Obtain all node IDs for layers in the tree");

        function obtainNodeIdForLayerId(layerId)
        {
            runCommand({
                command: "LayerTree.nodeIdForLayerId",
                parameters: {"layerId": layerId},
                callback: gotNodeIdForLayer
            });
        };

        layerIds.forEach(obtainNodeIdForLayerId);
    };

    var nodeIds = [];

    function gotNodeIdForLayer(result)
    {
        var id = result.nodeId;

        nodeIds.push(id);

        if (!--layerCount)
            finishTest();
    };

    function finishTest()
    {
        // We pass if we got this far and we have some node IDs reported to be 0,
        // which is what is expected to be returned for layers not associated with
        // a Node, for instance CSS generated content.
        if (nodeIds.indexOf(0) !== -1)
            InspectorTest.log("PASS");
        else
            InspectorTest.log("FAIL: Did not get any node ID equal to 0.");
        
        InspectorTest.completeTest();
    };

    function flattenedLayerTree(layerTree)
    {
        var layerByIds = {};

        function recurse(layer)
        {
            layerByIds[layer.layerId] = layer;
            if (layer.childLayers)
                layer.childLayers.forEach(recurse);
        };
        
        recurse(layerTree);

        return layerByIds;
    };

    function step(test)
    {
        logTestName(test.name);
        runCommand(test);
    };

    function logTestName(name)
    {
        InspectorTest.log("\n=== " + name + " ===\n");
    };

    function runCommand(command)
    {
        InspectorTest.sendCommand(command.command, command.parameters, function(messageObject) {
            if (messageObject.hasOwnProperty("error")) {
                InspectorTest.log("FAIL: " + messageObject.error.message + " (" + messageObject.error.code + ")");
                InspectorTest.completeTest();
                return;
            }

            if (command.name)
                InspectorTest.log("PASS");

            command.callback(messageObject.result);
        });
    };

};

window.addEventListener("DOMContentLoaded", function()
{
    runTest();
}, false);

</script>
<style type="text/css">
      
    div::after {
        position: absolute;
        width: 100px;
        height: 100px;
        background-color: black;
        -webkit-transform: translateZ(0);
        content: "";
    }

</style>
</head>
<body>

    <div></div>

</body>
</html>
