LayoutTests/inspector-protocol/layer-tree-generated-content-expected.txt


=== Get the Document ===

PASS

=== Enable the LayerTree agent ===

PASS

=== Get the layer tree ===

PASS

=== Obtain all node IDs for layers in the tree ===

PASS
